requirements.txt
asgiref==3.2.10
astroid==2.4.2
boto3==1.15.16
botocore==1.18.16
cachetools==4.1.1
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.13
django-heroku==0.3.1
django-allauth==0.54.0
django-autoslug==1.9.8
django-cleanup==5.1.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-filter==2.4.0
django-pwa==1.0.10
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-storages==1.10.1
djangorestframework==3.11.1
docutils==0.15.2
future==0.16.0
google-api-core==1.22.4
google-auth==1.22.1
google-cloud-core==1.4.3
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
isort==5.6.3
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauth2==1.9.0.post1
oauthlib==3.1.0
pathspec==0.5.9
protobuf==3.15.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
semantic-version==2.5.0
six==1.15.0
social-auth-core==3.3.3
sqlparse==0.4.2
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wcwidth==0.1.9
wrapt==1.12.1
whitenoise==4.1.2